+++ packages/bridge/bridge-react-webpack-plugin/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @module-federation/bridge-react-webpack-plugin
 
+## 0.18.4
+
+### Patch Changes
+
+- Updated dependencies [8061f8c]
+  - @module-federation/sdk@0.18.4
+
 ## 0.18.3
 
 ### Patch Changes